requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-session==0.5.0
firebase-admin==6.2.0
python-dotenv==1.0.0
PyPDF2==3.0.1
spacy==3.7.2
requests==2.31.0
google-generativeai==0.3.1
requests
gunicorn
